+++ NEWS.md
@@ -37,6 +37,7 @@ Enhancements:
   * New function `plotResidency()` can be used to plot the individual residency for each fish.
   * The column order in global.ratios now follows the section order.
   * New output in the residency analysis: group.ratios (similar to global.ratios, but per group).
+  * New function `plotRatios()` can be used to plot the global and group residency.
 
 ## actel 1.1.0